# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Método Ponto B - RE/MAX SC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals
      parameters:
        - name: franchiseId
          in: query
          schema:
            type: integer
        - name: dimensionId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get a goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalDetail"
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}/kpis:
    get:
      operationId: listGoalKpis
      tags: [kpis]
      summary: List KPIs for a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of KPIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kpi"
    post:
      operationId: createKpi
      tags: [kpis]
      summary: Create a KPI for a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
  /kpis/{id}:
    patch:
      operationId: updateKpi
      tags: [kpis]
      summary: Update a KPI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
    delete:
      operationId: deleteKpi
      tags: [kpis]
      summary: Delete a KPI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /goals/{id}/initiatives:
    get:
      operationId: listGoalInitiatives
      tags: [goalInitiatives]
      summary: List initiatives for a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of goal initiatives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalInitiative"
    post:
      operationId: createGoalInitiative
      tags: [goalInitiatives]
      summary: Add an initiative to a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInitiativeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalInitiative"
  /goal-initiatives/{id}:
    get:
      operationId: getGoalInitiative
      tags: [goalInitiatives]
      summary: Get a goal initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal initiative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalInitiative"
    patch:
      operationId: updateGoalInitiative
      tags: [goalInitiatives]
      summary: Update a goal initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInitiativeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalInitiative"
...
components:
  schemas:
...
    Goal:
      type: object
      required: [id, franchiseId, dimensionId, keyProcessId, title, status, progressPercentage, score, createdAt]
      properties:
        id:
          type: integer
        franchiseId:
          type: integer
        franchiseName:
          type: ["string", "null"]
        dimensionId:
          type: integer
        dimensionName:
          type: ["string", "null"]
        keyProcessId:
          type: integer
        keyProcessName:
          type: ["string", "null"]
        title:
          type: string
        kriDescription:
          type: ["string", "null"]
        currentValue:
          type: ["number", "null"]
        targetValue:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        ownerUserId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        frequency:
          type: ["string", "null"]
        status:
          type: string
          enum: [nao_iniciada, em_andamento, em_atencao, atrasada, concluida, pausada]
        progressPercentage:
          type: number
        riskStatus:
          type: ["string", "null"]
          enum: [atrasado, no_prazo, adiantado, null]
        score:
          type: number
        activeInitiativesCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    GoalDetail:
      type: object
      required: [id, franchiseId, dimensionId, keyProcessId, title, status, progressPercentage, score, createdAt]
      properties:
        id:
          type: integer
        franchiseId:
          type: integer
        franchiseName:
          type: ["string", "null"]
        dimensionId:
          type: integer
        dimensionName:
          type: ["string", "null"]
        keyProcessId:
          type: integer
        keyProcessName:
          type: ["string", "null"]
        title:
          type: string
        kriDescription:
          type: ["string", "null"]
        currentValue:
          type: ["number", "null"]
        targetValue:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        ownerUserId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        frequency:
          type: ["string", "null"]
        status:
          type: string
        progressPercentage:
          type: number
        riskStatus:
          type: ["string", "null"]
        score:
          type: number
        activeInitiativesCount:
          type: ["integer", "null"]
        createdAt:
          type: string
        kpis:
          type: array
          items:
            $ref: "#/components/schemas/Kpi"
        initiatives:
          type: array
          items:
            $ref: "#/components/schemas/GoalInitiative"
    GoalInput:
      type: object
      required: [franchiseId, dimensionId, keyProcessId, title]
      properties:
        franchiseId:
          type: integer
        dimensionId:
          type: integer
        keyProcessId:
          type: integer
        title:
          type: string
        kriDescription:
          type: string
        currentValue:
          type: number
        targetValue:
          type: number
        unit:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        ownerUserId:
          type: integer
        frequency:
          type: string
          enum: [diario, semanal, mensal, trimestral, semestral, anual]
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        kriDescription:
          type: string
        currentValue:
          type: number
        targetValue:
          type: number
        unit:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        ownerUserId:
          type: integer
        frequency:
          type: string
        status:
          type: string
          enum: [nao_iniciada, em_andamento, em_atencao, atrasada, concluida, pausada]
    Kpi:
      type: object
      required: [id, goalId, name, indicatorType, desiredDirection, createdAt]
      properties:
        id:
          type: integer
        goalId:
          type: integer
        name:
          type: string
        initialValue:
          type: ["number", "null"]
        currentValue:
          type: ["number", "null"]
        targetValue:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        frequency:
          type: ["string", "null"]
        indicatorType:
          type: string
          enum: [numero_absoluto, percentual, valor_financeiro, quantidade, sim_nao]
        desiredDirection:
          type: string
          enum: [aumentar, reduzir, manter]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    KpiInput:
      type: object
      required: [name, indicatorType, desiredDirection]
      properties:
        name:
          type: string
        initialValue:
          type: number
        currentValue:
          type: number
        targetValue:
          type: number
        unit:
          type: string
        frequency:
          type: string
          enum: [diario, semanal, mensal, trimestral, semestral, anual]
        indicatorType:
          type: string
          enum: [numero_absoluto, percentual, valor_financeiro, quantidade, sim_nao]
        desiredDirection:
          type: string
          enum: [aumentar, reduzir, manter]
        notes:
          type: string
    KpiUpdate:
      type: object
      properties:
        name:
          type: string
        currentValue:
          type: number
        targetValue:
          type: number
        unit:
          type: string
        frequency:
          type: string
        indicatorType:
          type: string
        desiredDirection:
          type: string
        notes:
          type: string
    GoalInitiative:
      type: object
      required: [id, goalId, strategicInitiativeId, status, progressPercentage, createdAt]
      properties:
        id:
          type: integer
        goalId:
          type: integer
        strategicInitiativeId:
          type: integer
        initiativeName:
          type: ["string", "null"]
        dimensionName:
          type: ["string", "null"]
        keyProcessName:
          type: ["string", "null"]
        desiredResult:
          type: ["string", "null"]
        actualResult:
          type: ["string", "null"]
        mainKpiId:
          type: ["integer", "null"]
        ownerUserId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        frequency:
          type: ["string", "null"]
        executionDay:
          type: ["string", "null"]
        executionTime:
          type: ["string", "null"]
        estimatedTime:
          type: ["string", "null"]
        whatWillBeDone:
          type: ["string", "null"]
        whyItMatters:
          type: ["string", "null"]
        whoIsResponsible:
          type: ["string", "null"]
        whereItWillBeDone:
          type: ["string", "null"]
        howItWillBeDone:
          type: ["string", "null"]
        investmentOrEffort:
          type: ["string", "null"]
        progressPercentage:
          type: number
        status:
          type: string
          enum: [ativa, concluida, pausada, cancelada]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...